feat: wire unit field into misc repository and models
- Add `unit` to `RecipeAdditionMisc::try_from` in models.rs
- Derive `amount_is_weight` from `unit` in `MiscRepository::create` (g/oz → true, others → false)
- Update `MiscRepository::update` to sync `amount_is_weight` when `unit` changes
- Propagate `unit` through recipe clone, version snapshot, and version restore paths
- Default `unit` in BeerXML import (derive from `amount_is_weight` field)
- Fix pre-existing `batch_attachments` i32/i64 type mismatch from entity regen

+++ src-tauri/src/repositories/batch_attachments.rs
@@ -33,7 +33,7 @@ impl<'a> BatchAttachmentRepository<'a> {
             filename: Set(filename.to_string()),
             original_name: Set(original_name.to_string()),
             mime_type: Set(mime_type.map(|s| s.to_string())),
-            size_bytes: Set(size_bytes),
+            size_bytes: Set(size_bytes as i32),
             created_at: Set(now),
         }
         .insert(self.db)
@@ -72,7 +72,7 @@ impl<'a> BatchAttachmentRepository<'a> {
             filename: m.filename,
             original_name: m.original_name,
             mime_type: m.mime_type,
-            size_bytes: m.size_bytes,
+            size_bytes: m.size_bytes as i64,
             created_at: m.created_at as i64,
         }
     }
